Programming Project #2 Update

+++ Programming Project #2/ExpenseAccount.java
@@ -57,7 +57,7 @@ public class ExpenseAccount {
         output.println("Invoice For: " + this.getAccountName());
         // Loop through file to determine all charges
         while (scan.hasNextLine()) {
-            double charge = scan.nextInt();
+            double charge = scan.nextDouble();
             try {
                 // Make payment and print to output
                 this.makePayment(charge);
